requirements.txt
fastapi
uvicorn
twilio
python-dotenv
requests
jinja2
python-multipart
httpx
pandas
openpyxl
sqlalchemy
openai
groq
